Cambiar propiedades Weight y Height de double a float
Se actualizaron las propiedades `Weight` y `Height` de la clase
`Pokemon` en el archivo `Pokemon.cs`, cambiando su tipo de
dato de `double` a `float`. Este cambio busca optimizar el
uso de memoria y se considera que la precisión de `float`
es suficiente para los valores manejados por estas
propiedades.

+++ src/Domain/Entities/Pokemon.cs
@@ -6,8 +6,8 @@ namespace Domain.Entities
     {
         public string? Name { get; set; }
         public string? Image { get; set; }
-        public double Weight { get; set; }
-        public double Height { get; set; }
+        public float Weight { get; set; }
+        public float Height { get; set; }
         public ICollection<string>? Types { get; set; }
         public ICollection<PokemonStat>? Stats { get; set; }
     }
